+++ kgcnn/md/ase_calc.py
@@ -1,4 +1,5 @@
 import ase
+import tensorflow as tf
 import numpy as np
 from ase import Atoms
 from ase.calculators.calculator import Calculator, all_changes
@@ -7,6 +8,8 @@ from kgcnn.graph.serial import get_preprocessor
 from kgcnn.data.base import MemoryGraphList
 from copy import deepcopy
 from typing import Union, List
+
+ks = tf.keras
 
 
 class AtomsToGraphConverter:
@@ -70,27 +73,39 @@ class AtomsToGraphConverter:
         return config
 
 
-class KgcnnCalculator(ase.calculators.calculator.Calculator):
+class EnergyForceExtensivePostProcessor:
+
+    def __init__(self, scaler: None):
+        self.scaler = scaler
+
+    def __call__(self, model_output_dict: dict, atoms):
+        energy, forces = model_output_dict["energy"], model_output_dict["forces"]
+        _, energy, forces = self.scaler(X=None, y=energy, force=forces)
+        model_output_dict.update({"energy": energy, "forces": forces})
+        return model_output_dict
+
+
+class KgcnnSingleCalculator(ase.calculators.calculator.Calculator):
     r"""ASE calculator for machine learning models from :obj:`kgcnn`."""
 
     def __init__(self,
-                 models: list = None,
+                 model: ks.models.Model = None,
                  model_inputs: Union[list, dict] = None,
                  model_outputs: Union[list, dict] = None,
-                 converter: AtomsToGraphConverter = None,
+                 atoms_converter: AtomsToGraphConverter = None,
                  graph_preprocessors: List[dict] = None,
-                 scaler: list = None,
+                 results_postprocessors=None,
                  **kwargs):
-        super(KgcnnCalculator, self).__init__(**kwargs)
+        super(KgcnnSingleCalculator, self).__init__(**kwargs)
 
-        self.converter = converter
-        self.models = models
+        self.atoms_converter = atoms_converter
+        self.model = model
         self.model_inputs = model_inputs
         self.model_outputs = model_outputs
         self.graph_preprocessors = graph_preprocessors
-        self.scaler = scaler
+        self.results_postprocessors = results_postprocessors
 
-    def _model_load(self, file_path: Union[List[str], str]) -> list:
+    def _model_load(self, file_path: str) -> list:
         pass
 
     @staticmethod
@@ -108,28 +123,25 @@ class KgcnnCalculator(ase.calculators.calculator.Calculator):
             raise TypeError("'%s' output translation must be 'str', 'dict' or 'list'." % properties)
         return output
 
-    def _model_predict(self, atoms: Union[List[Atoms], Atoms]):
-        graph_list = self.converter(atoms)  # type MemoryGraphList
+    def _model_predict(self, atoms: Union[List[Atoms], Atoms]) -> dict:
+        graph_list = self.atoms_converter(atoms)  # type MemoryGraphList
+
         graph_list.map_list(self.graph_preprocessors)
         tensor_input = graph_list.tensor(self.model_inputs)
-        outputs = {}
-        for s, m in zip(self.scaler, self.models):
+            
+        try:
+            tensor_output = self.model(tensor_input, training=False)
+        except ValueError:
+            tensor_output = self.model.predict(tensor_input)
 
-            try:
-                tensor_output = m(tensor_input, training=False)
-            except ValueError:
-                tensor_output = m.predict(tensor_input)
+        # Translate output
+        tensor_dict = self._translate_properties(tensor_output, self.model_outputs)
 
-            # Translate output
-            tensor_output = self._translate_properties(tensor_output, self.model_outputs)
+        # Cast to numpy
+        output_dict = {key: value.numpy() for key, value in tensor_dict.items()}
 
-            # Rescale output
-            if s is not None:
-                if hasattr(s, "_standardize_coordinates"):
-                    assert s._standardize_coordinates is False, "Scaling of model input is not supported."
-                tensor_output = s.inverse_transform(tensor_output)
+        return output_dict
 
-            # Prepare output.
 
     # Interface to ASE calculator scheme.
     def calculate(self, atoms=None, properties=None, system_changes=None):
@@ -137,3 +149,15 @@ class KgcnnCalculator(ase.calculators.calculator.Calculator):
         if not self.calculation_required(atoms, properties):
             # Nothing to do.
             return
+
+        output_dict = self._model_predict(atoms)
+
+        # Apply PostProcessor
+        if self.results_postprocessors:
+            if isinstance(self.results_postprocessors, (list, tuple)):
+                for mpp in self.results_postprocessors:
+                    output_dict = mpp(output_dict, atoms)
+            else:
+                output_dict = self.results_postprocessors(output_dict, atoms)
+
+        self.results = output_dict